requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
PyPDF2==3.0.1
python-docx==1.1.0
aiofiles==23.2.1
google-generativeai==0.3.2
openpyxl
